Refine elog

+++ include/linux/elogk.h
@@ -18,9 +18,9 @@
 struct eevent_t {
     __u16 len;             /* length of the payload */
     __u16 type;            /* system call number */
-    __s16 id;              /* ID for the entry, pos:invoke, neg:ret */
+    __s32 id;              /* ID for the entry, pos:invoke, neg:ret */
     /* identify the subject generating this log. maybe pid or uid */
-    __u16 belong;
+    __u32 belong;
     struct timespec etime;  /* timestamp for the entry */
     char payload[0];        /* the entry's payload (syscall params) */
 } __attribute__ ((packed));
